pyproject.toml
[project]
name = "interview-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.49.0",
    "pdfplumber>=0.11.7",
    "pypdf>=6.1.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "ruff>=0.14.4",
    "tenacity>=9.1.2",
]
